### 2021-01-11 #2950 @whedon
[PRE REVIEW]: uDALES: large-eddy-simulation software for urban flow, dispersion, and microclimate modelling
**Submitting author:** @tomgrylls (<a href="http://orcid.org/0000-0002-0948-4932">Tom Grylls</a>)
**Repository:** <a href="https://github.com/uDALES/u-dales" target ="_blank">https://github.com/uDALES/u-dales</a>
**Version:** 1.0.0
