+++ paper/paper.bib
@@ -278,12 +278,14 @@
   copyright = {Creative Commons Attribution NonCommercial NoDerivatives Licence}
 }
 
-
-@unpublished{Grylls2021,
-  author    = {Grylls, Tom},
-  publisher = {Elsevier {BV}},
-  journal   = {Agricultural and Forest Meteorology},
-  title     = {Tree model with drag, transpiration, shading and deposition: identification of cooling regimes and large-eddy simulation},
-  year      = {2021},
-  notes     = {under review (accepted)}
+@article{Grylls2021,
+  title = {Tree model with drag, transpiration, shading and deposition: Identification of cooling regimes and large-eddy simulation},
+  journal = {Agricultural and Forest Meteorology},
+  volume = {298-299},
+  pages = {108288},
+  year = {2021},
+  issn = {0168-1923},
+  doi = {https://doi.org/10.1016/j.agrformet.2020.108288},
+  url = {http://www.sciencedirect.com/science/article/pii/S0168192320303907},
+  author = {Tom Grylls and Maarten van Reeuwijk}
 }
