### 2021-01-13 b4446e1
Update author names
Edits to author names to fix formatting of surnames with multiple words.
