+++ paper/paper.bib
@@ -57,7 +57,7 @@
   volume    = {3},
   number    = {2},
   pages     = {415--444},
-  author    = {T. Heus and C. C. van Heerwaarden and H. J. J. Jonker and A. Pier Siebesma and S. Axelsen and K. van den Dries and O. Geoffroy and A. F. Moene and D. Pino and S. R. de Roode and J. Vil{\`{a}}-Guerau de Arellano},
+  author    = {T. Heus and C. C. {van Heerwaarden} and H. J. J. Jonker and A. Pier Siebesma and S. Axelsen and K. {van den Dries} and O. Geoffroy and A. F. Moene and D. Pino and S. R. de Roode and J. Vil{\`{a}}-Guerau {de Arellano}},
   title     = {Formulation of the Dutch Atmospheric Large-Eddy Simulation ({DALES}) and overview of its applications},
   journal   = {Geoscientific Model Development}
 }
@@ -81,7 +81,7 @@
   year      = {2020},
   month     = oct,
   publisher = {Springer Science and Business Media {LLC}},
-  author    = {Birgit S. S\"{u}tzl and Gabriel G. Rooney and Maarten van Reeuwijk},
+  author    = {Birgit S. S\"{u}tzl and Gabriel G. Rooney and Maarten {van Reeuwijk}},
   title     = {Drag Distribution in Idealized Heterogeneous Urban Environments},
   journal   = {Boundary-Layer Meteorology}
 }
@@ -187,7 +187,7 @@
   volume    = {39},
   number    = {6},
   pages     = {1281--1286},
-  author    = {V. Rammohan Rao and V.M.K. Sastri},
+  author    = {Rammohan V. Rao and V.M.K. Sastri},
   title     = {Efficient evaluation of diffuse view factors for radiation},
   journal   = {International Journal of Heat and Mass Transfer}
 }
@@ -229,7 +229,7 @@
   publisher = {American Geophysical Union ({AGU})},
   volume    = {12},
   number    = {8},
-  author    = {D. Meyer and R. Schoetter and M. Riechert and A. Verrelle and M. Tewari and J. Dudhia and V. Masson and M. Reeuwijk and S. Grimmond},
+  author    = {D. Meyer and R. Schoetter and M. Riechert and A. Verrelle and M. Tewari and J. Dudhia and V. Masson and M. {van Reeuwijk} and S. Grimmond},
   title     = {{WRF}-{TEB}: Implementation and Evaluation of the Coupled Weather Research and Forecasting ({WRF}) and Town Energy Balance ({TEB}) Model},
   journal   = {Journal of Advances in Modeling Earth Systems}
 }
@@ -256,7 +256,7 @@
   publisher = {Elsevier {BV}},
   volume    = {3},
   pages     = {100041},
-  author    = {Tom Grylls and Cl{\'{e}}mence M.A. Le Cornec and Pietro Salizzoni and Lionel Soulhac and Marc E.J. Stettler and Maarten van Reeuwijk},
+  author    = {Tom Grylls and Cl{\'{e}}mence M.A. Le Cornec and Pietro Salizzoni and Lionel Soulhac and Marc E.J. Stettler and Maarten {van Reeuwijk}},
   title     = {Evaluation of an operational air quality model using large-eddy simulation},
   journal   = {Atmospheric Environment: X}
 }
@@ -264,7 +264,7 @@
 @unpublished{Suter2021,
   year      = {2021},
   publisher = {Copernicus {GmbH}},
-  author    = {Ivo Suter and Tom Grylls and Birgit S. S\"{u}tzl and Maarten van Reeuwijk},
[…]
