### 2021-01-14 #2950 @Kevin-Mattheus-Moerman
@tomgrylls I'll be looking for an editor to handle this submission. In the mean time it would be great if you could help suggest reviewers for this work. Also can you please review our documentation on [what the paper should contain](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain), in particular the "Statement of need" section is missing. You can update the paper here by calling: `@whedon generate pdf`. Thanks.
