+++ paper/paper.bib
@@ -187,7 +187,7 @@
   volume    = {39},
   number    = {6},
   pages     = {1281--1286},
-  author    = {Rammohan V. Rao and V.M.K. Sastri},
+  author    = {V. {Rammohan Rao} and V.M.K. Sastri},
   title     = {Efficient evaluation of diffuse view factors for radiation},
   journal   = {International Journal of Heat and Mass Transfer}
 }
