### 2021-01-16 1ea674c
Add macOS and Debug to job matrix
