+++ tests/run_tests.py
@@ -108,6 +108,9 @@ def run_and_compare(cases_dir, path_to_exes, is_patch=False):
 
             # For driver sims we need to copy all files in first from the precursor simulation.
             if case_id in driver_sims:
+                if case_id == '502' and platform == "darwin": # FIXME: need to renable this.
+                    print('Skipping test for case 502 on macOS, see #131')
+                    return
                 for f_name in (model_output_dir.parents[1] / '501' / model_output_dir.name).glob('*driver*'):
                     shutil.copy(f_name, model_output_dir.parents[1] / '502' / model_output_dir.name)
 
@@ -121,15 +124,13 @@ def run_and_compare(cases_dir, path_to_exes, is_patch=False):
                                     model_output_dirs[0].parent)
 
 def run_udales(path_to_exe: Path, namelist: str, model_output_dir: str, 
-               model_output_dirs: list, cpu_count=None) -> None:
-    if cpu_count is None:
-        cpu_count = str(os.cpu_count())
+               model_output_dirs: list, cpu_count=2) -> None:
     print(f'Running uDALES in: {path_to_exe}')
     try:
-        subprocess.run(['mpiexec', '-np', cpu_count, path_to_exe / 'u-dales',
-                        namelist], cwd=model_output_dir, check=True, 
-                        stdout=subprocess.DEVNULL, stderr=subprocess.STDOUT)
-    except:
+        subprocess.run(['mpiexec', '-np', str(cpu_count), path_to_exe / 'u-dales',
+                        namelist], cwd=model_output_dir, check=True,
+                        stdout=subprocess.DEVNULL)
+    except subprocess.CalledProcessError:
         print(f'Could not run case uDALES in {path_to_exe} for namelist {namelist}')
         sys.exit(1)
     model_output_dirs.append(model_output_dir)
