+++ tests/run_tests.py
@@ -108,9 +108,6 @@ def run_and_compare(cases_dir, path_to_exes, is_patch=False):
 
             # For driver sims we need to copy all files in first from the precursor simulation.
             if case_id in driver_sims:
-                # if case_id == '502' and platform == "darwin": # FIXME: need to renable this.
-                #     print('Skipping test for case 502 on macOS, see #131')
-                #     return
                 for f_name in (model_output_dir.parents[1] / '501' / model_output_dir.name).glob('*driver*'):
                     shutil.copy(f_name, model_output_dir.parents[1] / '502' / model_output_dir.name)
 
@@ -124,7 +121,7 @@ def run_and_compare(cases_dir, path_to_exes, is_patch=False):
                                     model_output_dirs[0].parent)
 
 def run_udales(path_to_exe: Path, namelist: str, model_output_dir: str, 
-               model_output_dirs: list, cpu_count=4) -> None:
+               model_output_dirs: list, cpu_count=2) -> None:
     print(f'Running uDALES in: {path_to_exe}')
     try:
         subprocess.run(['mpiexec', '-np', str(cpu_count), path_to_exe / 'u-dales',
