### 2021-01-16 9e04609
temporarily disable 502 on macOS
