### 2021-01-17 #2950 @tomgrylls
@Kevin-Mattheus-Moerman<https://github.com/Kevin-Mattheus-Moerman> thanks for your email. If more editor suggestions would be helpful, I have been through the list and would suggest the following:

kyleniemeyer
wkearn
abhishekbajpayee
wkearn
mzamoraz

In terms of the statement of need, we have integrated this requirement into the main body of the text (from the end of the first paragraph) as we felt this helped the readability of the paper and thought that a section title would break the paper up. However, if having the statement of the need requirement within the text is insufficient then please let us know and we will add a separate section and heading.

Thanks,

Tom
