### 2021-01-21 #2950 @Kevin-Mattheus-Moerman
Thanks @wimvanderbauwhede that's great. We are hearing this from a lot of editors and reviewers at the moment so this is totally understandable. We usually aim for a 2 week review time but we can and have been very flexible especially in these challenging times. I'd say we will start the review within a week and the review will likely be ongoing for about 1-2 months (given peoples current time constraints) (although usually people submit minor initial comments right away given our open review process). Does that sound reasonable? Let me know what works for you.
