### 2021-01-21 #2950 @Kevin-Mattheus-Moerman
@leifdenby @jakubfabian @crshapiro @tonyinme @claireverhulst @JuhaTonttila @fjansson would you be interested in reviewing this work for [the Journal of Open Source Software (JOSS)](https://joss.theoj.org/): uDALES: large-eddy-simulation software for urban flow, dispersion, and microclimate modelling?

The JOSS [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) takes place on GitHub and focuses on [the software](https://github.com/uDALES/u-dales) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.02950/joss.02950/10.21105.joss.02950.pdf).  This issue is a "pre-review" issue in which reviewers are assigned. Once sufficient reviewers are recruited we will open a dedicated review issue where the review will take place. [Here](https://github.com/openjournals/joss-reviews/issues/2831) is an example an ongoing review process to give you an idea.

If you are interested in reviewing, you can comment here. Thanks!
