### 2021-01-21 #2950 @crshapiro
Thank you for inviting me to review software for this innovative journal. Unfortunately, I don't have the time to commit to being a reviewer for something like this.
