### 2021-02-01 071c048
Re-enable CI for macOS and Debug (#133)
