### 2021-02-08 #2950 @Kevin-Mattheus-Moerman
@leifdenby @jakubfabian @tonyinme @claireverhulst @JuhaTonttila @fjansson would you be interested in reviewing this work for the Journal of Open Source Software (JOSS): uDALES: large-eddy-simulation software for urban flow, dispersion, and microclimate modelling?

The JOSS review process takes place on GitHub and focuses on the software and a short paper. This issue is a "pre-review" issue in which reviewers are assigned. Once sufficient reviewers are recruited we will open a dedicated review issue where the review will take place. Here is an example an ongoing review process to give you an idea.

If you are interested in reviewing, you can comment here. Thanks!
