### 2021-02-08 #2950 @tonyinme
Hey Kevin,

Thanks for the invite.
Unfortunately, I don't have the time to commit to this.

Thanks,
Tony

On Mon, Feb 8, 2021 at 3:27 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @leifdenby <https://github.com/leifdenby> @jakubfabian
> <https://github.com/jakubfabian> @tonyinme <https://github.com/tonyinme>
> @claireverhulst <https://github.com/claireverhulst> @JuhaTonttila
> <https://github.com/JuhaTonttila> @fjansson <https://github.com/fjansson>
> would you be interested in reviewing this work for the Journal of Open
> Source Software (JOSS): uDALES: large-eddy-simulation software for urban
> flow, dispersion, and microclimate modelling?
>
> The JOSS review process takes place on GitHub and focuses on the software
> and a short paper. This issue is a "pre-review" issue in which reviewers
> are assigned. Once sufficient reviewers are recruited we will open a
> dedicated review issue where the review will take place. Here is an example
> an ongoing review process to give you an idea.
>
> If you are interested in reviewing, you can comment here. Thanks!
>
