### 2021-02-23 #3055 @wimvanderbauwhede
@Kevin-Mattheus-Moerman  Two questions:
1. Regarding opening issues on the project repository under review: is there some tag we should use to make it clear that this is an issue related to the JOSS review?
2. "Has the submitting author (@tomgrylls) made major contributions to the software?" In terms of LoC it is clearly "no" (0.2% of changes), all we can say is that he has several paper on the topic. What is the guidance on this?
