### 2021-02-23 #3055 @dmey
@wimvanderbauwhede and @Kevin-Mattheus-Moerman I appreciate that the list of authors as reported by the cloc utility may lead to some confusion so let me clarify this. Most commits made during development were squashed therefore the current metadata is not very indicative of the amount of work and contributions made by each author. Furthermore given that this project is a fork of DALES, our commits appear together with those from the DALES team (we state in the summary that uDALES is a fork of DALES). @tomgrylls and @ivosuter made key contributions in terms of scientific and feature development as reflected by the current list of authors. The order of co-authors has been discussed and agreed beforehand.
