### 2021-02-24 #3055 @p-costa
OK, I should mention something, @Kevin-Mattheus-Moerman -- I realize now that my former colleagues from the Aero and Hydrodynamics Laboratory at TU Delft (Netherlands), Jasper Thomas and Mathieu Pourquie (not co-authors of this paper) have contributed to this code. Nevertheless, I feel comfortable with reviewing this, and my review would surely be unbiased.
