### 2021-02-24 #3055 @zbeekman
@Kevin-Mattheus-Moerman Thanks for the offer! In my experience additional reviewers has never made for a quicker and more seamless review process. On the flip side, I'll informally offer any thoughts to improve the work and the authors can address them if they find it helpful.

This is my suggestion: What if I work on an informal review (I'll keep a copy of the reviewer checklist markdown privately) and add comments & suggestions that I think could be helpful. In the event another reviewer isn't able to complete their review I can step in and become an official reviewer (and I'll be at least part of the way through the official review checklist by that point).
