### 2021-02-28 275ee62
Fixes to bottom edge wall function as identified in #23
