+++ tests/run_tests.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 
 # uDALES (https://github.com/uDALES/u-dales).
-# Copyright (C) 2019 D. Meyer.
 
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -15,6 +14,8 @@
 
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+# Copyright (C) 2019 the uDALES Team.
 
 
 """Test uDALES
