+++ tools/singularity/udales_build.sh
@@ -1,4 +1,22 @@
 #!/usr/bin/env bash
+
+# uDALES (https://github.com/uDALES/u-dales).
+
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+# Copyright (C) 2016-2019 the uDALES Team.
+
 set -e
 
 # Usage: udales_build.sh <NPROC> [Debug, Release]
