+++ src/program.f90
@@ -1,19 +1,23 @@
 !> \file program.f90
 !! Main program
 
-!>
-!! \mainpage
-!! Dutch Atmospheric Large Eddy Simulation -URBAN
-!! \section DALES Dutch Atmospheric Large Eddy Simulation -URBAN
-!!
-!! @version 48
-!!
-!
-!! \todo
-!!
 !! \section License License
-!!  This file is part of DALESURBAN.
-!!  Copyright 1993-2014 Delft University of Technology
+!!  This file is part of DALES.
+!!
+!!  DALES is free software; you can redistribute it and/or modify it under the
+!! terms of the GNU General Public License as published by the Free Software
+!! Foundation; either version 3 of the License, or (at your option) any later
+!! version.
+!!
+!!  DALES is distributed in the hope that it will be useful, but WITHOUT ANY
+!! WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+!! PARTICULAR PURPOSE.  See the GNU General Public License for more details.
+!!
+!!  You should have received a copy of the GNU General Public License along with
+!! this program.  If not, see <http://www.gnu.org/licenses/>.
+!!
+!!  Copyright 1993-2009 Delft University of Technology, Wageningen University,
+!! Utrecht University, KNMI
 !!
 program DALESURBAN      !Version 48
 
