+++ src/modstatistics.f90
@@ -3,14 +3,14 @@
 !>
 !!  \author Tom Grylls, ICL Dec 16 2016
 !
-!  This file is part of DALES.
+!  This file is part of uDALES.
 !
-! DALES is free software; you can redistribute it and/or modify
+! This program is free software: you can redistribute it and/or modify
 ! it under the terms of the GNU General Public License as published by
-! the Free Software Foundation; either version 3 of the License, or
+! the Free Software Foundation, either version 3 of the License, or
 ! (at your option) any later version.
 !
-! DALES is distributed in the hope that it will be useful,
+! This program is distributed in the hope that it will be useful,
 ! but WITHOUT ANY WARRANTY; without even the implied warranty of
 ! MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 ! GNU General Public License for more details.
@@ -18,7 +18,7 @@
 ! You should have received a copy of the GNU General Public License
 ! along with this program.  If not, see <http://www.gnu.org/licenses/>.
 !
-!  Copyright 1993-2009 Delft University of Technology, Wageningen University, Utrecht University, KNMI
+!  Copyright 2006-2021 the uDALES Team.
 !
 module modstatistics
 
