+++ src/moddriver.f90
@@ -11,6 +11,24 @@
 !! \todo Documentation
 !!       Remove unecessary "use" variables
 !!       Remove unecessary commented lines
+!
+!  This file is part of uDALES.
+!
+! This program is free software: you can redistribute it and/or modify
+! it under the terms of the GNU General Public License as published by
+! the Free Software Foundation, either version 3 of the License, or
+! (at your option) any later version.
+!
+! This program is distributed in the hope that it will be useful,
+! but WITHOUT ANY WARRANTY; without even the implied warranty of
+! MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+! GNU General Public License for more details.
+!
+! You should have received a copy of the GNU General Public License
+! along with this program.  If not, see <http://www.gnu.org/licenses/>.
+!
+!  Copyright 2006-2021 the uDALES Team.
+!
 module moddriver
 use modinletdata
 implicit none

[…]
