### 2021-03-07 e580661
Add copyright notice to wall function modules
