### 2021-03-11 uDALES/u-dales#137 @wimvanderbauwhede
JOSS review comments on Review checklist requirements
The Review checklist at https://github.com/openjournals/joss-reviews/issues/3055 has a few items that I can't tick at the moment because I can't find the info in the docs, so if you could please point me to it:

"Documentation
 A statement of need: Does the paper have a section titled 'Statement of Need' that clearly states what problems the software is designed to solve and who the target audience is?"
> I could not find any section in the docs titled 'Statement of Need'

"Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?"
> There are examples, but there does not seem to be a "make test" or similar build feature to perform tests on the build.

"Functionality
 [...]
 Functionality: Have the functional claims of the software been confirmed?
 Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)"
> I could not find anything in the docs about validation of functionality or performance.
