### 2021-03-11 uDALES/u-dales#138 @wimvanderbauwhede
JOSS review: comments on installation and running
I went through the Getting Started guide and managed to install and run everything fine, but here are some comments on the installation and running to help you improve the documentation.

The "Getting Started" guide  does not tell me where to get the code.

So I cloned the master branch
```
https://github.com/uDALES/u-dales.git
```
## Singularity

Singularity-based installation requires golang and some other packages.  The instructions to install golang are messy, I just used the Ubuntu packages and that worked fine.

"To build and download the Singularity image use:"

	singularity build --remote tools/singularity/image.sif tools/singularity/image.def

	FATAL:   Unable to submit build job: no authentication token, log in with `singularity remote login`

So I did

	singularity remote login

It asks me to create a token and paste it

I did this, I got:

	Access Token:

		FATAL:   while verifying token: error response from server: Invalid Credentials

What worked:

	singularity remote login --tokenfile ./sylabs-token

So once again I did

	singularity build --remote tools/singularity/image.sif tools/singularity/image.def

That takes a while. Some indication of how long it may take would be welcome

"then, to install uDALES use:"

	./tools/singularity/udales_build.sh 2 Release

Ditto

"Finally, to run an example case use:"

	./tools/singularity/udales_run.sh 2 Release examples/001 namoptions.001

That works but an explanation of how to inspect the results would be very useful.

# Prerequisites

That should maybe say "Prerequisites when not using Singularity"

The Linux/WSL (Ubuntu) instructions work fine but could have come earlier because I had already followed the link to the netcdf web site, downloaded netcdf-c and netcdf-fortran and was about to start building when I thought, there must be a better way.
