### 2021-03-11 uDALES/u-dales#137 @dmey
1. Statement of need
I think that this had already been addressed in https://github.com/openjournals/joss-reviews/issues/2950#issuecomment-761789660 but let us know if this not the case.

2. Automated Tests
We have a series of automated regression and compile tests that are run at every commit -- https://github.com/uDALES/u-dales/tree/master/tests

3. Functionality
I am not sure what is meant by this exactly -- I do not think there are any specific claims made that need to be checked but @tomgrylls may have more to say about this.
