### 2021-03-11 uDALES/u-dales#138 @dmey
Thanks for the feedback! I'd be happy to create a PR that addresses how to clone and clarify the prerequisites section for cases when singularity is not used. With regards to the installation of Singularity I am not sure -- as this is a separate dependency, currently we assume that information on how to install and set up Singularity are sourced directly from Singularity's website (dependencies and instructions may change at any time) there how would you feel if instead of adding specific doc information about this we were to link to the Singularity's website and ask the user to refer to that for up to date information about the installation and use of Singularity?

With regards to "an explanation of how to inspect the results would be very useful", we have a section at the end of the page (https://udales.github.io/u-dales/udales-getting-started/#whats-next) that takes you to simulate and view the results -- did you find this too hidden away? We wanted to avoid duplication but maybe there is a better place to add this information. @tomgrylls @samoliverowens any thoughts/preferences?
