### 2021-03-11 uDALES/u-dales#137 @wimvanderbauwhede
1. The form distinguishes between Statement of Need in the Documentation and in the Software Paper. In the paper I am fine with the way it is addressed in  https://github.com/openjournals/joss-reviews/issues/3055 but apparently you need this in the docs as well
2. Great, but I would not have found them had you not pointed it out. Could this maybe be added to the guide?
3. The paper claims "capable of explicitly representing buildings and of modelling the full range of urban processes". I am sure you explain this in detail in your papers, but as  JOSS reviewer I am expected to learn this from the documentation. If there are no specific claims about performance then that is fine of course.
