### 2021-03-11 uDALES/u-dales#137 @dmey
1. @tomgrylls I think this needs to be added to docs as well -- maybe by adding an overview section in the README.md that is a cutout from the current paper?
2. @wimvanderbauwhede if you don't mind I think this is a bit out of scope -- this is linked from the development guide which is itself linked from the main README.md. Software testing is less of a end-user facing task.
3.  @tomgrylls this is something more on your side
