### 2021-03-11 uDALES/u-dales#138 @wimvanderbauwhede
Re "an explanation of how to inspect the results would be very useful": I found a way. I searched the web for a netcdf viewer and found Paraview and Panoply. Panoply works just fine to visualise the results.
