### 2021-03-11 f2ce610
add info about how to inspect the results
