### 2021-03-12 uDALES/u-dales#138 @samoliverowens
ncview can also be used to visualise the results quickly.
