### 2021-03-12 uDALES/u-dales#138 @dmey
> ncview can also be used to visualise the results quickly.

👍. Added in 0d36037
