### 2021-03-12 4d048be
Update docs (closes #138) (#139)
