+++ udales-docs-software/src/modstatistics.f90
@@ -1,0 +1,400 @@
+!> \file modstatistics.f90
+!!  Calculates field statistics to be written in modstatsdump.f90
+!>
+!!  \author Tom Grylls, ICL Dec 16 2016
+!
+!  This file is part of DALES.
+!
+! DALES is free software; you can redistribute it and/or modify
+! it under the terms of the GNU General Public License as published by
+! the Free Software Foundation; either version 3 of the License, or
+! (at your option) any later version.
+!
+! DALES is distributed in the hope that it will be useful,
+! but WITHOUT ANY WARRANTY; without even the implied warranty of
+! MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+! GNU General Public License for more details.
+!
+! You should have received a copy of the GNU General Public License
+! along with this program.  If not, see <http://www.gnu.org/licenses/>.
+!
+!  Copyright 1993-2009 Delft University of Technology, Wageningen University, Utrecht University, KNMI
+!
+module modstatistics
+
+  use modglobal, only : dt,ltkedump
+  use modmpi, only : myid
+  implicit none
+  private
+  PUBLIC :: genstats,tkestats
+  save
+
+  !NetCDF variables
+  integer :: klow,khigh,i,j,k
+!  real    :: tsamplep,tstatsdumpp,tsample,tstatsdump
+
+contains
+
+  !-------------------------
+  !> Calculate general stats
+  !-------------------------
+
+  subroutine genstats(tsamplep,tstatsdumpp,umint,vmint,wmint)
+
+  use modfields,        only : um,up,vm,wm,thlm,uav,vav,wav,uuav,vvav,wwav,uvav,vwav,uwav,thlav,thlwav,thlthlav, &
+                               upupav,vpvpav,wpwpav,upvpav,upwpav,vpwpav,thlpwpav
+  use modglobal,        only : ib,ie,ih,jb,je,dy,jh,ke,kb,kh,rk3step,timee,cexpnr,tsample,tstatsdump,&
+                               ltempeq,dxf,dzf,dzhi
+  use modmpi,           only : myid,cmyid,my_real,mpi_sum,mpierr,comm3d
+  implicit none
+
+  real, dimension(ib-ih:ie+ih,jb-jh:je+jh,kb:ke+kh)     :: umint
+  real, dimension(ib-ih:ie+ih,jb-jh:je+jh,kb:ke+kh)     :: vmint
+  real, dimension(ib-ih:ie+ih,jb-jh:je+jh,kb:ke+kh)     :: wmint
+  real :: tstatsdumppi,tsamplep,tstatsdumpp
+  integer :: km
+
+  tstatsdumppi = 1./tstatsdumpp
+
+!  if (lydump) then
+    if (.not. rk3step==3)  return
+!      if (tsamplep > tsample) then
+
+        !> Interpolate velocity fields to cell centers
+!        do k=kb-kh,ke
+!          do j=jb-jh,je+jh
+!            do i=ib-ih,ie+ih
+!              umint(i,j,k) = 0.5*(um(i,j,k)+um(i+1,j,k))
+!              vmint(i,j,k) = 0.5*(vm(i,j,k)+vm(i,j+1,k))
+!              wmint(i,j,k) = 0.5*(wm(i,j,k)+wm(i,j,k+1))
+!            enddo
[…]
