+++ udales-docs-software/src/moddriver.f90
@@ -1,0 +1,720 @@
+!! moddriverinlet.f90 contains the set up for creating a driver simulation to use an inlet conditions for a following simulation/
+!! This creates a realistic turbulent inlet profiles as necessary.
+!! The fields are extracted from a specified plane and stored in
+!! respective files. These files are then used to create inlet
+!! conditions for a following simualtion. Fields are linearly
+!! interpolated in time where necessary.
+!! Code set up is adapted from modinlet.f90 since similar processes are involved
+
+!! \author Anton Esmail-Yakas, Imperial College London, August 5th 2017.
+!! Edited by tg3315, ICL, May 2019.
+!! \todo Documentation
+!!       Remove unecessary "use" variables
+!!       Remove unecessary commented lines
+module moddriver
+use modinletdata
+implicit none
+save
+  public :: initdriver,exitdriver,readdriverfile,drivergen
+
+contains
+  subroutine initdriver
+    use modglobal, only : ih,ib,ie,jh,jb,je,kb,ke,kh,jhc,khc,idriver,iplane,xf,lstoreplane,nstore,Uinf,ltempeq,lmoist,pi,zf,zh,driverstore,tdriverstart,tdriverdump,timeleft,dtdriver,nsv,timee
+    use modfields, only : um
+    use modmpi, only : myid,nprocs
+
+    implicit none
+    real    :: pfi, epsi
+    integer :: k
+
+    if (idriver==1) then
+      if (tdriverstart < timee) then
+        write(0, *) 'ERROR: tdriverstart must be greater than the elapsed time at the start of the simulation'
+        stop 1
+      end if
+      tdriverdump = tdriverstart
+   endif
+
+    if (idriver==1) then
+      allocate(storetdriver(1:driverstore))
+      allocate(storeu0driver(jb-jh:je+jh,kb-kh:ke+kh,1:driverstore))
+      allocate(storeumdriver(jb-jh:je+jh,kb-kh:ke+kh,1:driverstore))
+      allocate(storev0driver(jb-jh:je+jh,kb-kh:ke+kh,1:driverstore))
+      allocate(storevmdriver(jb-jh:je+jh,kb-kh:ke+kh,1:driverstore))
+      allocate(storew0driver(jb-jh:je+jh,kb-kh:ke+kh,1:driverstore))
+      allocate(storewmdriver(jb-jh:je+jh,kb-kh:ke+kh,1:driverstore))
+      !allocate(storee120driver(jb-jh:je+jh,kb-kh:ke+kh,1:driverstore))
+      !allocate(storee12mdriver(jb-jh:je+jh,kb-kh:ke+kh,1:driverstore))
+      if (ltempeq  ) then
+        allocate(storethl0driver(jb-jh:je+jh,kb-kh:ke+kh,1:driverstore))
+        allocate(storethlmdriver(jb-jh:je+jh,kb-kh:ke+kh,1:driverstore))
+      end if
+      if (lmoist  ) then
+        allocate(storeqt0driver(jb-jh:je+jh,kb-kh:ke+kh,1:driverstore))
+        allocate(storeqtmdriver(jb-jh:je+jh,kb-kh:ke+kh,1:driverstore))
+      end if
+      if (nsv>0  ) then
+        allocate(storesv0driver(jb-jhc:je+jhc,kb-khc:ke+khc,1:nsv,1:driverstore))
+        allocate(storesvmdriver(jb-jhc:je+jhc,kb-khc:ke+khc,1:nsv,1:driverstore))
+      end if
+
[…]
