+++ udales-docs-software/src/modmpi.f90
@@ -1,0 +1,432 @@
+!> \file modmpi.f90
+!!  Layer to deal with the parallelization.
+
+!>
+!!  Layer to deal with the parallelization.
+!>
+!!  \author Matthieu Pourquie, TU Delft
+!!  \par Revision list
+!!  \todo Documentation
+!!  \todo 2D/3D parallelization
+!!  \todo Include interfaces for MPI_ALLREDUCE, MPI_ALLTOALL, MPI_BCAST,
+!! MPI_SENDRECV to get rid of pure mpi calls in the code
+!  This file is part of DALES.
+!
+! DALES is free software; you can redistribute it and/or modify
+! it under the terms of the GNU General Public License as published by
+! the Free Software Foundation; either version 3 of the License, or
+! (at your option) any later version.
+!
+! DALES is distributed in the hope that it will be useful,
+! but WITHOUT ANY WARRANTY; without even the implied warranty of
+! MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+! GNU General Public License for more details.
+!
+! You should have received a copy of the GNU General Public License
+! along with this program.  If not, see <http://www.gnu.org/licenses/>.
+!
+!  Copyright 1993-2009 Delft University of Technology, Wageningen University, Utrecht University, KNMI
+!
+
+
+
+module modmpi
+use mpi
+implicit none
+save
+  integer comm3d
+  integer nbrtop
+  integer nbrbottom
+  integer myid
+  integer nprocs
+  integer mpierr
+  integer my_real
+  real    CPU_program    !end time
+  real    CPU_program0   !start time
+
+   character(3) :: cmyid
+
+contains
+  subroutine initmpi
+    implicit none
+    integer dims(1)
+    logical periods(1)
+    integer periods2(1)
+
+    call MPI_INIT(mpierr)
+    MY_REAL = MPI_DOUBLE_PRECISION  !MPI_REAL8 should be the same..
+    call MPI_COMM_RANK( MPI_COMM_WORLD, myid, mpierr )
+    call MPI_COMM_SIZE( MPI_COMM_WORLD, nprocs, mpierr )
+! Specify the # procs in each direction.
+! specifying a 0 means that MPI will try to find a useful # procs in
+! the corresponding  direction,
+
+! specifying 1 means only 1 processor in this direction, meaning that
+! we have in fact a grid of (at most) 2 dimensions left. This is used
+! when we want the array index range in 1 particular direction to be
+! present on all processors in the grid
+
+    dims(1) = 0
+
+
+! directions 1 and 2 are chosen periodic
+
+
+    periods(1) = .true.
+! Soares 20080115
+    periods2(1) = 1
+
+! find suitable # procs in each direction
+
+    call MPI_DIMS_CREATE( nprocs, 1, dims, mpierr )
+
+! create the Cartesian communicator, denoted by the integer comm3d
+
+    ! BUG - Thijs, Harm
+    !call MPI_CART_CREATE(MPI_COMM_WORLD, 1, dims, periods,.false., &
+    !                    comm3d, ierr )
+
[…]

[…]
