### 2021-03-14 4155faf
Add overview section to README based off JOSS paper
