### 2021-03-14 a829109
Initial text for validation section
