+++ src/modstartup.f90
@@ -13,6 +13,7 @@
 !!  \todo documentation
 !!  \par Revision list
 !
+!  This file is part of DALES.
 !
 ! DALES is free software; you can redistribute it and/or modify
 ! it under the terms of the GNU General Public License as published by
