+++ src/modsubgriddata.f90
@@ -12,6 +12,8 @@
 !!  \par Revision list
 !!  \todo Documentation
 !
+!  This file is part of DALES.
+!
 ! DALES is free software; you can redistribute it and/or modify
 ! it under the terms of the GNU General Public License as published by
 ! the Free Software Foundation; either version 3 of the License, or
