+++ tools/examples/plot_fielddump_slice.m
@@ -1,3 +1,20 @@
+% uDALES (https://github.com/uDALES/u-dales).
+
+% This program is free software: you can redistribute it and/or modify
+% it under the terms of the GNU General Public License as published by
+% the Free Software Foundation, either version 3 of the License, or
+% (at your option) any later version.
+
+% This program is distributed in the hope that it will be useful,
+% but WITHOUT ANY WARRANTY; without even the implied warranty of
+% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+% GNU General Public License for more details.
+
+% You should have received a copy of the GNU General Public License
+% along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+% Copyright (C) 2016-2021 the uDALES Team.
+
 function plot_fielddump_slice(expnr, field_var, slice_var, slice_id, time_id)
 
 % Plots a slice of fielddump in 2D and 3D.
