+++ README.md
@@ -11,13 +11,25 @@ This is the development repository for the uDALES (urban Dutch Atmospheric Large
 
 With continuing urbanization, challenges associated with the urban environment such as air quality, heat islands, pedestrian thermal comfort, and wind loads on tall buildings, are increasingly relevant. Our ability to realistically capture processes such as the transport of heat, moisture, momentum and pollutants, and those of radiative transfer in urban environments is key to understanding and facing these challenges.
 
-The turbulent nature of the urban flow field and the inherent heterogeneity and wide range of scales associated with the urban environment result in a complex modelling problem. Large-eddy simulation (LES) is an approach to turbulence modelling used in computational fluid dynamics to simulate turbulent flows over a wide range of spatial and temporal scales. LES is one of the most promising tools to model the interactions typical of urban areas due to its ability to resolve the urban flow field at resolutions of _O_(1 m, 0.1 s), over spatial domains of _O_(100 m), and time periods of _O_(1 h). Although there are many scalable LES models for atmospheric flows, to our knowledge, only few are capable of explicitly representing buildings and of modelling the full range of urban processes.
+The turbulent nature of the urban flow field and the inherent heterogeneity and wide range of scales associated with the urban environment result in a complex modelling problem. Large-eddy simulation (LES) is an approach to turbulence modelling used in computational fluid dynamics to simulate turbulent flows over a wide range of spatial and temporal scales. LES is one of the most promising tools to model the interactions typical of urban areas due to its ability to resolve the urban flow field at resolutions of _O_(1 m, 0.1 s), over spatial domains of _O_(100 m), and time periods of _O_(1 h). Although there are many scalable LES models for atmospheric flows, to our knowledge, only few are capable of explicitly representing buildings and of modelling the full range of urban processes (e.g. PALM-4U; Resler et al. 2017; Maronga et al. 2020; or OpenFoam; Weller et al. 1998)..
 
 uDALES is capable of modelling urban environments at the microclimate scale including wet thermodynamics, idealized and complex morphologies, three-dimensional urban surface energy balance and real-time pollution dispersion.
 
-## Validation
+## Publications and validation
 
-The dynamic core, DALES, has been validated extensively and used as part of several atmospheric intercomparison studies over a period of 30 years (Nieuwstadt and Brost, 1986; Heus et al., 2010). The use of the immersed boundary method to model obstacles has been validated against both wind and water tunnel data, for idealised geometries by Tomas et al. 2015 and complex geometries by Grylls, 2020. The wall functions and three-dimensional surface energy balance model was validated by Suter, 2018. The ability to model pollutuon dispersion in cities has been validated by Grylls et al., 2020 and Grylls, 2020. 
+To-date, uDALES has been used for a number of research applications:
+
+Table: uDALES research applications.
+
+| Research application                                         | Reference                                                                       |
+| ------------------------------------------------------------ | ------------------------------------------------------------------------------- |
+| Urban boundary layers/ boundary-layer meteorology            | Grylls et al. (2019); Sutzl et al. (2020)                                       |
+| Urban climate (radiation, green roofs and walls, trees etc.) | Suter (2018); Suter et al. (2021); Grylls & van Reeuwijk (2021)                 |
[…]
