### 2021-03-18 e919940
Minor textual fixes
