### 2021-03-18 a159b6e
Minor textual changes #2
