### 2021-03-18 uDALES/u-dales#137 @tomgrylls
Thanks both - I have added to the documentation and created a PR (#140).

@dmey I will first request a review from you and another uDALES team member to make sure we are happy with these additions before then confirming that it satisfies the checklist as discussed above.
