### 2021-03-20 97c1026
Updates from David and Birgit's reviews
