+++ paper/paper.bib
@@ -75,15 +75,16 @@
   journal   = {Atmospheric Environment}
 }
 
-@article{Sutzl2020,
-  doi       = {10.1007/s10546-020-00567-0},
-  url       = {https://doi.org/10.1007/s10546-020-00567-0},
-  year      = {2020},
-  month     = oct,
-  publisher = {Springer Science and Business Media {LLC}},
-  author    = {Birgit S. S\"{u}tzl and Gabriel G. Rooney and Maarten {van Reeuwijk}},
-  title     = {Drag Distribution in Idealized Heterogeneous Urban Environments},
-  journal   = {Boundary-Layer Meteorology}
+@article{Sutzl2021,
+  doi={10.1007/s10546-020-00567-0},
+  title={Drag Distribution in Idealized Heterogeneous Urban Environments},
+  author={S{\"u}tzl, Birgit S and Rooney, Gabriel G and {van Reeuwijk}, Maarten},
+  journal={Boundary-Layer Meteorology},
+  volume={178},
+  number={2},
+  pages={225--248},
+  year={2021},
+  publisher={Springer}
 }
 
 @article{Pourquie2009,
@@ -276,6 +277,18 @@
   publisher = {Imperial College London},
   year      = {2020},
   copyright = {Creative Commons Attribution NonCommercial NoDerivatives Licence}
+}
+
+@article{Grylls2020b,
+  doi={https://doi.org/10.1007/s10546-020-00508-x},
+  title={Steady-state large-eddy simulations of convective and stable urban boundary layers},
+  author={Grylls, Tom and Suter, Ivo and {van Reeuwijk}, Maarten},
+  journal={Boundary-Layer Meteorology},
+  volume={175},
+  number={3},
+  pages={309--341},
+  year={2020},
+  publisher={Springer}
 }
 
 @article{Grylls2021,
