+++ paper/paper.md
@@ -41,7 +41,7 @@ bibliography: paper.bib
 
 With continuing urbanization, challenges associated with the urban environment such as air quality, heat islands, pedestrian thermal comfort, and wind loads on tall buildings, are increasingly relevant. Our ability to realistically capture processes such as the transport of heat, moisture, momentum and pollutants, and those of radiative transfer in urban environments is key to understanding and facing these challenges [@Oke2017]. The turbulent nature of the urban flow field and the inherent heterogeneity and wide range of scales associated with the urban environment result in a complex modelling problem. Large-eddy simulation (LES) is an approach to turbulence modelling used in computational fluid dynamics to simulate turbulent flows over a wide range of spatial and temporal scales. LES is one of the most promising tools to model the interactions typical of urban areas due to its ability to resolve the urban flow field at resolutions of $O$(1 m, 0.1 s), over spatial domains of $O$(100 m), and time periods of $O$(1 h). Although there are many scalable LES models for atmospheric flows, to our knowledge, only few are capable of explicitly representing buildings and of modelling the full range of urban processes (e.g. PALM-4U @Resler2017; @Maronga2020; or OpenFoam @Weller1998).
 
-uDALES (urban Dutch Atmospheric LES) is an extension of DALES (Dutch Atmospheric LES; @Heus2010, @Tomas2015), capable of modelling urban environments at the microclimate scale with wet thermodynamics (Table 1). The uDALES framework includes tools to enable users to model a wide variety of idealized and complex urban morphologies [@Sutzl2020]. uDALES uses an Arakawa C-grid and typically uses second-order central-differencing schemes. For scalar quantities, e.g. for pollution concentration, it is possible to use a kappa-scheme for advection to ensure positivity. A third-order Runge-Kutta time integration scheme is applied. The immersed boundary method is used to represent buildings [@Pourquie2009] and wall functions to calculate the surface scalar and momentum fluxes at the rough immersed boundaries [@Cai2011; @Cai2012; @Uno1995], based on a local Richardson number. Additional factors are considered to obtain the evapotranspiration fluxes from vegetated surfaces. The code uses Fast-Fourier transforms to efficiently solve the Poisson equation for pressure and is fully parallelized using MPI.
+uDALES (urban Dutch Atmospheric LES) is an extension of DALES (Dutch Atmospheric LES; @Heus2010, @Tomas2015), capable of modelling urban environments at the microclimate scale with wet thermodynamics (Table 1). The uDALES framework includes tools to enable users to model a wide variety of idealized and complex urban morphologies [@Sutzl2021]. uDALES uses an Arakawa C-grid and typically uses second-order central-differencing schemes. For scalar quantities, e.g. for pollution concentration, it is possible to use a kappa-scheme for advection to ensure positivity. A third-order Runge-Kutta time integration scheme is applied. The immersed boundary method is used to represent buildings [@Pourquie2009] and wall functions to calculate the surface scalar and momentum fluxes at the rough immersed boundaries [@Cai2011; @Cai2012; @Uno1995], based on a local Richardson number. Additional factors are considered to obtain the evapotranspiration fluxes from vegetated surfaces. The code uses Fast-Fourier transforms to efficiently solve the Poisson equation for pressure and is fully parallelized using MPI.
 
[…]
