### 2021-03-21 #3055 @p-costa
Hi @tomgrylls ,

I just read the article and it looks great. I have no major comments. Just two minor things based on the last proof (https://github.com/openjournals/joss-reviews/issues/3055#issuecomment-803294401):

1. Line 35 -- F in *Fast* should not be capitalized and the dash can be removed.
2. Line 128 -- Bibliographic details of the PhD thesis are missing?

Now, I have a question regarding the acknowledgement of other people who may have contributed to **u**DALES (i.e. the fork of DALES), who are not co-authors. Perhaps I'm wrong but I think that Jasper Tomas, Mathieu Pourquie, and maybe Harmen Jonker, have contributed to it by adding e.g. an immersed boundary and changing the Poisson solver?

These contributions are acknowledged in the source code, but I feel they should be made more explicit because the average user will never see them (they are also not in the commit history). In short, I wonder if there are other important contributors to uDALES who should be mentioned e.g. in a list of contributors on the root of the repo, acknowledged in the paper, etc, so that their work gets credited?
