### 2021-03-21 uDALES/u-dales#138 @ashwinvis
For the record, I had similar experience as @wimvanderbauwhede with the `singularity build --remote ...` instruction. I wonder if the `remote` option is required. I managed to get the build go forward by skipping it and executing  `sudo singularity build ...`.
