### 2021-03-21 #3055 @ashwinvis
Hi,
Some questions and comments forto you.@tomgrylls  before I check off my review lists.

## Notes

Corresponding to the line numbers from the latest manuscript

- 19: Substantiate the claim of typical spatial and temporal scales involved in an urban LES. Smallest scale of O(1 m) seems quite large to me. Also, O(1 h) seems too less as it will not capture a single diurnal cycle. See Figure 3 https://doi.org/10.5194/gmd-13-1335-2020 for instance.

- 24: Explicitly mention the key feature which differentiates uDALES and DALES. I had to find it from Sützl et al. (2020) which says, uDALES was extended to provide "the capability to model buildings using an immersed boundary method". Also a "statement of need" section is missing from the manuscript.

- 30: Reference for kappa-scheme is missing.

- 36: "fully parallelized". Mention what kind of domain decomposition is used. Slabs or pencils? Which directions are distributed over MPI ranks?

- 44: "study real-time pollution dispersion". It is unclear to me what real-time means here.

- 74: URL missing for Grylls (2020)

- 130: URL / DOI missing for Suter et al. (2021). Is it submitted and pending review? I am curious about this article as it is closely related to this manuscript. Do you have a preprint available?

- 136:  Tomas et al. (2015) mentions

> The main modifications are the addition of an immersed boundary method (Pourquie et al. (2009)), the implementation of inflow/outflow boundary conditions and the application of the eddy-viscosity subgrid model of Vreman (2004).

Here I have the same concern as @p-costa. It appears that the authors of the 2015 article contributed to the development of uDALES. If so, why are they not acknowledged here?

## On documentation

- Installation instruction can be improved
- Tests instructions (now in a standalone README.md) should be included in the main documentation
- Contributing guidelines can be included in the main documentation too.
