### 2021-03-22 uDALES/u-dales#138 @dmey
@ashwinvis thanks for the feedback -- we had updated the docs at https://github.com/uDALES/u-dales/pull/139. With regards to the specifics of Singularity, we ask users to refer to the Singularity docs as this is not something related to uDALES. With regards to the remote build option, we wanted something that works on most systems and that is what the remote options does. As you pointed out, if you can run a program with elevated privileges, you don't have to use the remote build but, in most HPC/shared environments, this is often not the case.
