### 2021-03-22 #3055 @dmey
@ashwinvis thanks for your feedback -- I can address your docs comments and will leave the others to @tomgrylls.

> Installation instruction can be improved

Can you be a bit more specific? What exactly needs to be improved? We have recently address the @wimvanderbauwhede's docs related comments in https://github.com/uDALES/u-dales/issues/138.

>Tests instructions (now in a standalone README.md) should be included in the main documentation

Tests instructions are linked from the development guide which is itself linked from the main README.md. Software testing is less of a end-user facing task. This was reported by @wimvanderbauwhede in https://github.com/uDALES/u-dales/issues/137. I don't see a good reason for adding it to the main README (will make it unnecessarily bloated) but please let me know if you/others think otherwise.

> Contributing guidelines can be included in the main documentation too.

Same as above -- we kept it separate and linked from the main README to keep it lean.
