### 2021-03-22 uDALES/u-dales#138 @ashwinvis
Fair enough. I also tried the normal `cmake` and `make` route today, and it did not go through. I ran it on ArchLinux, and here are

<summary>dependencies</summary>

```
community/netcdf-fortran 4.5.3-1 [installed]
    NetCDF fortran bindings
community/netcdf-openmpi 4.7.4-1 [installed]
    network Common Data Form interface for array-oriented data access and corresponding library with parallel support (openmpi version)
extra/openmpi 4.0.5-2 [installed]
    High performance message passing library (MPI)
local/nco 4.9.2-2 [installed]
    netCDF Operators allow users to manipulate and analyse data stored in NetCDF files
core/gcc-fortran 10.2.0-4 [installed: 10.2.0-6]
    usr/bin/gfortran
extra/cmake 3.19.7-1 [installed]
    A cross-platform open-source make system
extra/python 3.9.2-1 [installed]
    Next generation of the python high-level scripting language
extra/python-pip 20.3.1-1 [installed]
     The PyPA recommended tool for installing Python packages
```

<summary>build log</summary>

```fortran
❯ make
[  1%] Performing update step for 'fishpack-cmake-project'
[  3%] No patch step for 'fishpack-cmake-project'
[  5%] Performing configure step for 'fishpack-cmake-project'
-- Configuring done
-- Generating done
-- Build files have been written to: /home/avmo/src/sandbox/u-dales/build/release/external/fishpack-cmake-project-prefix/src/fishpack-cmake-project-build
[  7%] Performing build step for 'fishpack-cmake-project'
[ 92%] Built target objlib
[ 96%] Built target fishpack_static
[100%] Built target fishpack_shared
[  9%] No install step for 'fishpack-cmake-project'
[ 11%] Completed 'fishpack-cmake-project'
[ 15%] Built target fishpack-cmake-project
[ 17%] Performing update step for 'vfftpack-cmake-project'
[ 19%] No patch step for 'vfftpack-cmake-project'
[ 21%] Performing configure step for 'vfftpack-cmake-project'
-- Configuring done
-- Generating done
-- Build files have been written to: /home/avmo/src/sandbox/u-dales/build/release/external/vfftpack-cmake-project-prefix/src/vfftpack-cmake-project-build
[ 23%] Performing build step for 'vfftpack-cmake-project'
[ 93%] Built target objlib
[ 96%] Built target vfftpack_static
[100%] Built target vfftpack_shared
[ 25%] No install step for 'vfftpack-cmake-project'
[ 26%] Completed 'vfftpack-cmake-project'
[ 30%] Built target vfftpack-cmake-project
[ 32%] Building Fortran object CMakeFiles/u-dales.dir/src/initfac.f90.o
/home/avmo/src/sandbox/u-dales/src/initfac.f90:212:23:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  212 |         call MPI_BCAST(walltypes, nwallprops*nwalltypes, MY_REAL, 0, comm3d, mpierr)
      |                       1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:353:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  353 |          call MPI_BCAST(block, 11*nblocks, MPI_INTEGER, 0, comm3d, mpierr)
      |                        1
Error: Rank mismatch between actual argument at (1) and actual argument at (2) (scalar and rank-2)
/home/avmo/src/sandbox/u-dales/src/initfac.f90:355:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  355 |          call MPI_BCAST(faclGR(0:nfcts), nfcts + 1, mpi_logical, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (LOGICAL(4)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:356:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  356 |          call MPI_BCAST(facz0(0:nfcts), nfcts + 1, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:357:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  357 |          call MPI_BCAST(facz0h(0:nfcts), nfcts + 1, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:358:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  358 |          call MPI_BCAST(facalb(0:nfcts), nfcts + 1, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:359:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  359 |          call MPI_BCAST(facem(0:nfcts), nfcts + 1, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:360:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  360 |          call MPI_BCAST(facd(0:nfcts,1:nwalllayers),(nfcts+1)*nwalllayers, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:361:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  361 |          call MPI_BCAST(facdi(0:nfcts, 1:nwalllayers), (nfcts + 1)*nwalllayers, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:362:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  362 |          call MPI_BCAST(faccp(0:nfcts, 1:nwalllayers), (nfcts + 1)*nwalllayers, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:363:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  363 |          call MPI_BCAST(faclami(0:nfcts, 1:nwalllayers), (nfcts + 1)*nwalllayers, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:364:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  364 |          call MPI_BCAST(fackappa(0:nfcts, 1:nwalllayers+1), (nfcts + 1)*(nwalllayers+1), MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:365:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  365 |          call MPI_BCAST(faca(0:nfcts), nfcts + 1, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:366:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  366 |          call MPI_BCAST(facain(0:nfcts), nfcts + 1, MPI_Integer, 0, comm3d, mpierr)
      |                        1
Error: Rank mismatch between actual argument at (1) and actual argument at (2) (scalar and rank-1)
/home/avmo/src/sandbox/u-dales/src/initfac.f90:367:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  367 |          call MPI_BCAST(facets, 4*nfcts, MPI_Integer, 0, comm3d, mpierr)
      |                        1
Error: Rank mismatch between actual argument at (1) and actual argument at (2) (scalar and rank-2)
/home/avmo/src/sandbox/u-dales/src/initfac.f90:370:26:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  370 |            call MPI_BCAST(svf(1:nfcts), nfcts, MY_REAL, 0, comm3d, mpierr)
      |                          1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:371:26:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  371 |            call MPI_BCAST(netsw(1:nfcts), nfcts, MY_REAL, 0, comm3d, mpierr)
      |                          1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:375:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  375 |          call MPI_BCAST(Tfacinit(1:nfcts), nfcts, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:376:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  376 |          call MPI_BCAST(facT(0:nfcts, 1:nwalllayers+1), (nfcts + 1)*(nwalllayers+1), MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:377:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  377 |          call MPI_BCAST(facTdash(1:nfcts, 1:nwalllayers+1), (nfcts)*(nwalllayers+1), MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:378:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  378 |          call MPI_BCAST(facef(1:nfcts), nfcts, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:379:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  379 |          call MPI_BCAST(facefi(1:nfcts), nfcts, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:380:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  380 |          call MPI_BCAST(facefsum(1:nfcts), nfcts, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:381:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  381 |          call MPI_BCAST(fachf(0:nfcts), nfcts + 1, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:382:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  382 |          call MPI_BCAST(fachfi(0:nfcts), nfcts + 1, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:383:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  383 |          call MPI_BCAST(fachfsum(1:nfcts), nfcts, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:387:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  387 |          call MPI_BCAST(facf(0:nfcts, 1:5), (nfcts + 1)*5, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:388:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  388 |          call MPI_BCAST(fachurel(0:nfcts), nfcts + 1, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:389:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  389 |          call MPI_BCAST(facwsoil(0:nfcts), nfcts + 1, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:390:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  390 |          call MPI_BCAST(faccth(0:nfcts), nfcts + 1, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
/home/avmo/src/sandbox/u-dales/src/initfac.f90:394:24:

  195 |       call MPI_BCAST(nwalltypes, 1, MPI_Integer, 0, comm3d, mpierr)
      |                     2
......
  394 |          call MPI_BCAST(facqsat(0:nfcts), nfcts + 1, MY_REAL, 0, comm3d, mpierr)
      |                        1
Error: Type mismatch between actual argument at (1) and actual argument at (2) (REAL(8)/INTEGER(4)).
make[2]: *** [CMakeFiles/u-dales.dir/build.make:134: CMakeFiles/u-dales.dir/src/initfac.f90.o] Error 1
make[1]: *** [CMakeFiles/Makefile2:118: CMakeFiles/u-dales.dir/all] Error 2
make: *** [Makefile:103: all] Error 2

```

Do you understand this?
