### 2021-03-22 uDALES/u-dales#138 @dmey
Yes, we only support  GNU compilers versions 9 and below -- you seem to be running  10.2.
