### 2021-03-22 #3055 @ashwinvis
I will follow up on installation issues in uDALES/u-dales#138.

> Software testing is less of a end-user facing task.

Since these are research software very often the line between a user and a developer is very fuzzy. So it is a matter of taste if you want to have good on-boarding or not.

> which is itself linked from the main README.md

As for testing and contributing notes, it is definitely linked in the README, but the README is not included in the main documentation https://udales.github.io/u-dales/. Right now if I search for either in the documentation search bar, I get nothing, which may lead to wrong impression for the reader.  For now, I think a link to tests and contributing guidelines would be appreciated.
