### 2021-03-27 7a46e5e
Changes to README for JOSS review (#140)
