### 2021-03-28 a96190e
Add description of supported geometries to JOSS paper
