### 2021-03-28 a41483b
Update order of time periods for uDALES simulations
