### 2021-03-28 d725aa9
Add ref for kappa scheme
