+++ paper/paper.bib
@@ -7,6 +7,17 @@
   title     = {Urban Climates}
 }
 
+@article{Hundsdorfer1995,
+  title={A positive finite-difference advection scheme},
+  author={Hundsdorfer, Willem and Koren, Barry and Verwer, JG and others},
+  journal={Journal of computational physics},
+  volume={117},
+  number={1},
+  pages={35--46},
+  year={1995},
+  publisher={Elsevier}
+}
+
 @article{Resler2017,
   doi       = {10.5194/gmd-10-3635-2017},
   url       = {https://doi.org/10.5194/gmd-10-3635-2017},
