### 2021-03-28 44fb877
Update thesis references
