+++ paper/paper.bib
@@ -158,6 +158,8 @@
   author    = {Suter,  Ivo},
   title     = {Simulating the impact of blue-green infrastructure on the microclimate of urban areas},
   publisher = {Imperial College London},
+  school    = {Imperial College London},
+  note      = {PhD thesis},
   year      = {2018},
   copyright = {Creative Commons Attribution NonCommercial NoDerivatives Licence}
 }
@@ -287,6 +289,9 @@
   title     = {Simulating air pollution in the urban microclimate},
   publisher = {Imperial College London},
   year      = {2020},
+  note      = {(under embargo)},
+  type      = {PhD thesis},
+  school    = {Imperial College London},
   copyright = {Creative Commons Attribution NonCommercial NoDerivatives Licence}
 }
 
