+++ paper/paper.md
@@ -45,7 +45,7 @@ uDALES (urban Dutch Atmospheric LES) is an extension of DALES (Dutch Atmospheric
 
 A novel surface energy balance model has been implemented in a two-way-coupled manner [@Suter2018] and includes the effect of turbulent exchange of heat between the surface and the air, as well as radiation and thermal conduction within the surface. uDALES has tools for modelling shortwave and longwave radiative fluxes following a radiosity approach (similar to @Aoyagi2011; @Resler2017), including calculation of direct solar radiation and view factors [@RammohanRao1996], considering shading and multiple reflections. 
 
-uDALES also has the tools necessary to study real-time pollution dispersion within cities. Both idealized (point and line) and realistic (street network) sources can be implemented and both passive and reactive scalars can be modelled. The code supports null cycle chemistry (NO-NO$_2$-O$_3$ reactions), and can be easily extended to more sophisticated schemes [@Grylls2019; @Grylls2020].
+uDALES also has the tools necessary to study air quality within cities. Both idealized (point and line) and realistic (street network) sources can be implemented and both passive and reactive scalars can be modelled. The code supports null cycle chemistry (NO-NO$_2$-O$_3$ reactions), and can be easily extended to more sophisticated schemes [@Grylls2019; @Grylls2020]. The high spatial and temporal resolution enables analysis of e.g. real-time pedestrian pollution exposure.
 
 The modelling capabilities of uDALES outlined above, combined with the ability to devise numerous simulation set-ups (e.g. via different lateral boundary conditions: periodic, inflow-outflow and driver) facilitate a plethora of possible studies into the urban environment. Relevant recent publications and their research applications are summarized in Table 1. 
 
