+++ paper/paper.bib
@@ -281,7 +281,7 @@
   author    = {Ivo Suter and Tom Grylls and Birgit S. S\"{u}tzl and Maarten {van Reeuwijk}},
   title     = {{uDALES} 1.0.0: a large-eddy-simulation model for urban environments},
   journal   = {Geoscientific Model Development},
-  notes     = {in preparation}
+  note      = {in preparation}
 }
 
 @article{Grylls2020,
