### 2021-03-29 9fb2d8a
add thesis reference and fix bib typos
