+++ paper/paper.bib
@@ -68,7 +68,7 @@
   volume    = {3},
   number    = {2},
   pages     = {415--444},
-  author    = {T. Heus and C. C. {van Heerwaarden} and H. J. J. Jonker and A. Pier Siebesma and S. Axelsen and K. {van den Dries} and O. Geoffroy and A. F. Moene and D. Pino and S. R. de Roode and J. Vil{\`{a}}-Guerau {de Arellano}},
+  author    = {T. Heus and C. C. {van Heerwaarden} and H. J. J. Jonker and A. Pier Siebesma and S. Axelsen and K. {van den Dries} and O. Geoffroy and A. F. Moene and D. Pino and S. R. {de Roode} and J. Vil{\`{a}}-Guerau {de Arellano}},
   title     = {Formulation of the Dutch Atmospheric Large-Eddy Simulation ({DALES}) and overview of its applications},
   journal   = {Geoscientific Model Development}
 }
@@ -89,7 +89,7 @@
 @article{Sutzl2021,
   doi={10.1007/s10546-020-00567-0},
   title={Drag Distribution in Idealized Heterogeneous Urban Environments},
-  author={S{\"u}tzl, Birgit S and Rooney, Gabriel G and {van Reeuwijk}, Maarten},
+  author={S{\"u}tzl, Birgit S. and Rooney, Gabriel G. and {van Reeuwijk}, Maarten},
   journal={Boundary-Layer Meteorology},
   volume={178},
   number={2},
@@ -120,7 +120,7 @@
   volume    = {76},
   number    = {4},
   pages     = {395--409},
-  author    = {I. Uno and X -M. Cai and D. G. Steyn and S. Emori},
+  author    = {I. Uno and X.-M. Cai and D. G. Steyn and S. Emori},
   title     = {A simple extension of the Louis method for rough surface layer modelling},
   journal   = {Boundary-Layer Meteorology}
 }
@@ -133,7 +133,7 @@
   publisher = {Elsevier {BV}},
   volume    = {51},
   pages     = {268--277},
-  author    = {Xiaoming Cai},
+  author    = {Xiao-Ming Cai},
   title     = {Effects of differential wall heating in street canyons on dispersion and ventilation characteristics of a passive scalar},
   journal   = {Atmospheric Environment}
 }
@@ -317,4 +317,15 @@
   doi = {https://doi.org/10.1016/j.agrformet.2020.108288},
   url = {http://www.sciencedirect.com/science/article/pii/S0168192320303907},
   author = {Tom Grylls and Maarten {van Reeuwijk}}
+}
+
+@article{Sutzl2021thesis,
+  author    = {S{\"u}tzl, Birgit S.},
+  title     = {Rising from the ground: Distributed drag parameterization of urban environments for numerical weather prediction},
+  publisher = {Imperial College London},
+  year      = {2021},
+  note      = {(under embargo)},
+  type      = {PhD thesis},
+  school    = {Imperial College London},
+  copyright = {Creative Commons Attribution Licence}
 }
