### 2021-04-25 #3055 @tomgrylls
Thanks @p-costa for your questions, apologies for the delay in getting back to you on this. Please see our response below.

> 1. Line 35 -- F in _Fast_ should not be capitalized and the dash can be removed.

Thanks for pointing out - now amended

> 2. Line 128 -- Bibliographic details of the PhD thesis are missing?
>

These have now been updated. A couple of the PhD theses are still under embargo following completion. I have indicated where this is the case in the bibliography.

> Now, I have a question regarding the acknowledgement of other people who may have contributed to **u**DALES (i.e. the fork of DALES), who are not co-authors. Perhaps I'm wrong but I think that Jasper Tomas, Mathieu Pourquie, and maybe Harmen Jonker, have contributed to it by adding e.g. an immersed boundary and changing the Poisson solver?
>
> These contributions are acknowledged in the source code, but I feel they should be made more explicit because the average user will never see them (they are also not in the commit history). In short, I wonder if there are other important contributors to uDALES who should be mentioned e.g. in a list of contributors on the root of the repo, acknowledged in the paper, etc, so that their work gets credited?

Thanks for pointing this out. We have updated both the paper and README files to better acknowledge their contributions. Specifically, changed the text in the paper:

'The immersed boundary method, first introduced into DALES by @Pourquie2009 and @Tomas2015, is used to represent buildings (supporting grid-conforming cuboid geometries; @Pourquie2009).'

Update to the README, including a new acknowledgements section:

'The immersed boundary method, which is used to model obstacles in the fluid domain, was implemented into DALES and validated by Pourquie et al. (2009) and Tomas et al. (2015).'

'## Acknowledgements

The dynamic core of uDALES derives from the Dutch Atmospheric Large-Eddy Simulation ([DALES](https://github.com/dalesteam/dales)) model. We would like to acknowledge the work of the contributors to DALES over the years (Heus et al. 2010) and specifically the work of J. Tomas, M. Pourquie and H. Jonker in implementing the immersed boundary method into DALES (Tomas et al. 2015).'

> **EDIT**: In addition, I feel it would be good to mention in the paper/docs the type of geometries that can be accommodated as an immersed-boundary using the _stress IBM_, since this is a key feature of the code. The geometry needs to conform to the underlying grid, so I think that only cuboids or tetrahedrons are supported?

Thanks for pointing this out. I have updated the manuscript to make this clearer. I have also added in some more information in the uDALES [set-up documentation](https://github.com/uDALES/u-dales/blob/joss-paper/docs/udales-simulation-setup.md#buildings-and-blocks) and a reference to this on the documentation for example 201 where it advises on pre-defined buildings.

As you mention, the implementation of the IBM is such that the buildings must conform to the underlying grid and therefore only cuboids are currently supported in uDALES. The advantage of this implementation is that it is second-order accurate, the grid is relatively simple to generate and the pressure field can be solved via a fast Fourier-based transform method (the latter provides a big performance advantage). We have validated the use of grid conforming buildings for both idealised and realistic urban morphologies with good results. We are planning to update the IBM implementation in the future to capture more complex forms. As mentioned, the grid-conforming limitation has now been made clearer in the paper and documentation.
