### 2021-04-25 #3055 @tomgrylls
Thanks @ashwinvis for your questions, apologies for the delay in getting back to you on this. Please see our response below.

> * 19: Substantiate the claim of typical spatial and temporal scales involved in an urban LES. Smallest scale of O(1 m) seems quite large to me. Also, O(1 h) seems too less as it will not capture a single diurnal cycle. See Figure 3 https://doi.org/10.5194/gmd-13-1335-2020 for instance.

The resolution of O(1 m) is representative of the order of magnitude of resolution used across the current publications that have used uDALES to date. The range used across these is ~0.25-2 m. The use of wall functions means that refinement of the grid close to solid-fluid boundaries, as done in other urban LES, is not necessary. In terms of capturing typical urban morphologies, studies have shown that a minimum of 16 cells are required per building to capture flows in the urban canopy layer, with a resolution of 1 m this corresponds to building heights of approximately 16 m or larger, which is reasonable.

Agreed that that the choice of O(1 h) here is a misrepresentation of the general application of the code. For urban studies typically 10s of hours are needed to obtain statistically steady results. I have updated this to be 'O(10 h)', which is consistent with many of the simulations in the papers published using uDALES to date. Simulations of both O(1 h) and O(100 h) are also feasible using uDALES.

> * 24: Explicitly mention the key feature which differentiates uDALES and DALES. I had to find it from Sützl et al. (2020) which says, uDALES was extended to provide "the capability to model buildings using an immersed boundary method". Also a "statement of need" section is missing from the manuscript.

I have updated the manuscript to explicitly make this point at the start of the second paragraph (see lines 25-27). The remainder of the manuscript mentions the other features that have been added to the code in more detail.

The requirement for a statement of need was discussed with the editor and reviewer twice previously [here](https://github.com/uDALES/u-dales/issues/137#issuecomment-796725980), where they agreed that an additional section was not necessary as the statement of need was clear from the main body of text in the manuscript. Are you also okay with the statement of need being made within the text (see lines 9-23)?

> * 30: Reference for kappa-scheme is missing.

Thanks for highlighting - reference added.

> * 36: "fully parallelized". Mention what kind of domain decomposition is used. Slabs or pencils? Which directions are distributed over MPI ranks?

Updated the document to make this clear. Domain decomposition is performed in the spanwise direction, thereby producing slabs.

> * 44: "study real-time pollution dispersion". It is unclear to me what real-time means here.

Real-time in this sense referred to the fact that uDALES time-resolves the turbulent flow field and therefore the dispersion of pollutants unlike Reynolds-averaged Navier-Stokes (RANS) models that provide time averaged concentrations. It is therefore possible to study e.g. pedestrian exposure to pollution in 'real time'.

Agreed that this terminology was not clear. 'real time pollution dispersion' has been replaced by 'air quality' more generally (this is then inclusive of pollution emissions, dispersion, reaction and deposition). A sentence has been added to highlight the ability of the model to capture 'real time pedestrian exposure'.

> * 74: URL missing for Grylls (2020)

This thesis was placed under an embargo for 18 months following completion. The thesis will be made available in July 2021. I have added additional detail to the reference list and put a note of "(under embargo)" to capture this point.

> * 130: URL / DOI missing for Suter et al. (2021). Is it submitted and pending review? I am curious about this article as it is closely related to this manuscript. Do you have a preprint available?

This paper is still in preparation - I have added a note of (in preparation) to the reference list. The paper outlines the scientific details of the developments made in uDALES in more detail and also presents further validations of the code. Our hope is that that paper alongside this one will provide a complete overview of the uDALES software.

We do not yet have a preprint available to share. It is likely we will have a preprint available within the next 3 months once the paper has been finalised and submitted.

> * 136:  Tomas et al. (2015) mentions
>
> > The main modifications are the addition of an immersed boundary method (Pourquie et al. (2009)), the implementation of inflow/outflow boundary conditions and the application of the eddy-viscosity subgrid model of Vreman (2004).
>
> Here I have the same concern as @p-costa. It appears that the authors of the 2015 article contributed to the development of uDALES. If so, why are they not acknowledged here?
>
Replied in detail to the same question [here](https://github.com/openjournals/joss-reviews/issues/3055#issuecomment-826327758)
