### 2021-04-25 #3055 @p-costa
Thank you @tomgrylls for the update, and for clearly addressing all my queries 👍. Just one small note:

>'## Acknowledgements
> The dynamic core of uDALES derives from the Dutch Atmospheric Large-Eddy Simulation (DALES) model. We would like to acknowledge the work of the contributors to DALES over the years (Heus et al. 2010) and specifically the work of J. Tomas, M. Pourquie and H. Jonker in implementing the immersed boundary method into DALES (Tomas et al. 2015).'

This is fine, if you consider that the contributions of Tomas et al. deserve to be highlighted with respect to the previous ones. I would not want that they are highlighted simply because those where the contributions that I mentioned (and that I was specifically aware of). Else, the general acknowledgement seems better. No need to come back to me on this regard, please have it as you deem best.

Also, as a curiosity, perhaps you may be aware that this IBM can also accommodate slightly more complex shapes (e.g., in 2D, triangles). Although it may be tricker to have the grid conforming to them in this case. (See, e.g.,  [Benschop, H. O. G., and W-P. Breugem. "Drag reduction by herringbone riblet texture in direct numerical simulations of turbulent channel flow." Journal of Turbulence 18.8 (2017): 717-759.]; the article is behind paywall).
