### 2021-05-04 #3055 @tomgrylls
Thanks @p-costa - just confirming that it does makes sense to specifically acknowledge the names mentioned above as they are the only ones that have specifically contributed towards the development of uDALES from uDALES (in their work on immersed boundary methods in DALES), other contributions to DALES are less directly associated with uDALES.

Thanks for sharing the IBM reference - I will take a look!
