+++ paper/paper.bib
@@ -9,6 +9,7 @@
 
 @article{Hundsdorfer1995,
   title={A positive finite-difference advection scheme},
+  url={https://doi.org/10.1006/jcph.1995.1042},
   author={Hundsdorfer, Willem and Koren, Barry and Verwer, JG and others},
   journal={Journal of computational physics},
   volume={117},
@@ -281,7 +282,7 @@
   author    = {Ivo Suter and Tom Grylls and Birgit S. S\"{u}tzl and Maarten {van Reeuwijk}},
   title     = {{uDALES} 1.0.0: a large-eddy-simulation model for urban environments},
   journal   = {Geoscientific Model Development},
-  note      = {in preparation}
+  type      = {in preparation}
 }
 
 @article{Grylls2020,
