### 2021-05-04 #3055 @tomgrylls
> I am happy with the reply and the changes. I have some minor comments regarding the references and it may be taken up in the proofreading round.
>
> > > * 30: Reference for kappa-scheme is missing.
> >
> >
> > Thanks for highlighting - reference added.
>
> Please add a DOI / URL to keep the references consistent.
>

Now added

> > > * 74: URL missing for Grylls (2020)
> >
> >
> > This thesis was placed under an embargo for 18 months following completion. The thesis will be made available in July 2021. I have added additional detail to the reference list and put a note of "(under embargo)" to capture this point.
>
> Is it possible to have URL pointing to yours and Sützl's thesis so that it would be findable later on?
>

Good idea - I am not sure about if this is possible but I have enquired to Imperial College London to see if we can do this.

> > > * 130: URL / DOI missing for Suter et al. (2021). Is it submitted and pending review? I am curious about this article as it is closely related to this manuscript. Do you have a preprint available?
> >
> >
> > This paper is still in preparation - I have added a note of (in preparation) to the reference list. The paper outlines the scientific details of the developments made in uDALES in more detail and also presents further validations of the code. Our hope is that that paper alongside this one will provide a complete overview of the uDALES software.
>
> OK, I understand. I do not see the _(in preparation)_ appearing in the reference list.

Now added.

Thanks all again for your input on this!
