### 2021-05-09 #3055 @tomgrylls
@ashwinvis just confirming that I have now asked Imperial College London library and they are unable to mint DOIs while the thesis is still under embargo, so sadly it will not be possible to add DOIs to those 2 references at this time. Thanks again.
