### 2021-06-05 #3055 @Kevin-Mattheus-Moerman
@wimvanderbauwhede, @ashwinvis thanks for reviewing this work. I see all boxes are ticked and general positive comments. Can you confirm you are happy to recommend acceptance in JOSS at this point?
