### 2021-06-05 #3055 @Kevin-Mattheus-Moerman
@tomgrylls once I hear back from the reviewers I'll process acceptance shortly. If you'd like to proofread the paper once more yourself, now would be a good point. Also can you confirm if the version tag needs to be updated or if we are still at `1.0.0`?
