### 2021-06-05 #3055 @ashwinvis
@Kevin-Mattheus-Moerman I confirm that the paper can be accepted into JOSS
