### 2021-06-21 #3055 @Kevin-Mattheus-Moerman
@wimvanderbauwhede it looks like all your boxes are ticked, and all issues appear resolved. Can you confirm you are happy to recommend this work is published in JOSS?

To avoid further delay, and given the overall positive review by the reviewers, I will process this paper for acceptance this week, unless I hear objections from @wimvanderbauwhede.
