### 2021-06-29 #3055 @Kevin-Mattheus-Moerman
@tomgrylls congrats I will now continue to help get this submission ready for acceptance in JOSS. I have proofread the paper and it looks good. Please now work on the following steps:

- [ ] Since this work is about to be accepted I recommend that you also proofread the article once more.
- [ ] Can you confirm if the software version is still at `1.0.0` or if we need to update the version tag?
- [ ] Can you archive a copy of the reviewed software on ZENODO and report back here with the archived version DOI? You can archive this software manually or automatically (see these steps which some find useful: https://guides.github.com/activities/citable-code/).
- [ ] Make sure that the meta-data on ZENODO matches that of the paper. Especially check that the **title** and that **the set of authors**, the **author order** is the exact same. Note you can also link/add Orcid profiles to authors at ZENODO. You may need to edit this meta-data manually to ensure it matches the paper.
