### 2021-07-12 fb21d02
fix dois and urls
