+++ paper/paper.bib
@@ -89,6 +89,7 @@
 
 @article{Sutzl2021,
   doi={10.1007/s10546-020-00567-0},
+  url={https://doi.org/10.1007/s10546-020-00567-0},
   title={Drag Distribution in Idealized Heterogeneous Urban Environments},
   author={S{\"u}tzl, Birgit S. and Rooney, Gabriel G. and {van Reeuwijk}, Maarten},
   journal={Boundary-Layer Meteorology},
@@ -297,7 +298,8 @@
 }
 
 @article{Grylls2020b,
-  doi={https://doi.org/10.1007/s10546-020-00508-x},
+  doi={10.1007/s10546-020-00508-x},
+  url={https://doi.org/10.1007/s10546-020-00508-x},
   title={Steady-state large-eddy simulations of convective and stable urban boundary layers},
   author={Grylls, Tom and Suter, Ivo and {van Reeuwijk}, Maarten},
   journal={Boundary-Layer Meteorology},
@@ -315,8 +317,8 @@
   pages = {108288},
   year = {2021},
   issn = {0168-1923},
-  doi = {https://doi.org/10.1016/j.agrformet.2020.108288},
-  url = {http://www.sciencedirect.com/science/article/pii/S0168192320303907},
+  doi = {10.1016/j.agrformet.2020.108288},
+  url = {https://doi.org/10.1016/j.agrformet.2020.108288},
   author = {Tom Grylls and Maarten {van Reeuwijk}}
 }
 
