+++ paper/paper.bib
@@ -1,0 +1,334 @@
+@book{Oke2017,
+  doi       = {10.1017/9781139016476},
+  url       = {https://doi.org/10.1017/9781139016476},
+  year      = {2017},
+  publisher = {Cambridge University Press},
+  author    = {T. R. Oke and G. Mills and A. Christen and J. A. Voogt},
+  title     = {Urban Climates}
+}
+
+@article{Hundsdorfer1995,
+  title={A positive finite-difference advection scheme},
+  url={https://doi.org/10.1006/jcph.1995.1042},
+  author={Hundsdorfer, Willem and Koren, Barry and Verwer, JG and others},
+  journal={Journal of computational physics},
+  volume={117},
+  number={1},
+  pages={35--46},
+  year={1995},
+  publisher={Elsevier}
+}
+
+@article{Resler2017,
+  doi       = {10.5194/gmd-10-3635-2017},
+  url       = {https://doi.org/10.5194/gmd-10-3635-2017},
+  year      = {2017},
+  month     = oct,
+  publisher = {Copernicus {GmbH}},
+  volume    = {10},
+  number    = {10},
+  pages     = {3635--3659},
+  author    = {Jaroslav Resler and Pavel Kr{\v{c}} and Michal Belda and Pavel Juru{\v{s}} and Nina Bene{\v{s}}ov{\'{a}} and Jan Lopata and Ond{\v{r}}ej Vl{\v{c}}ek and Da{\v{s}}a Dama{\v{s}}kov{\'{a}} and Kry{\v{s}}tof Eben and P{\v{r}}emysl Derbek and Bj\"{o}rn Maronga and Farah Kanani-S\"{u}hring},
+  title     = {{PALM}-{USM} v1.0: A new urban surface model integrated into the {PALM} large-eddy simulation model},
+  journal   = {Geoscientific Model Development}
+}
+
+@article{Maronga2020,
+  doi       = {10.5194/gmd-13-1335-2020},
+  url       = {https://doi.org/10.5194/gmd-13-1335-2020},
+  year      = {2020},
+  month     = mar,
+  publisher = {Copernicus {GmbH}},
+  volume    = {13},
+  number    = {3},
+  pages     = {1335--1372},
+  author    = {Bj\"{o}rn Maronga and Sabine Banzhaf and Cornelia Burmeister and Thomas Esch and Renate Forkel and Dominik Fr\"{o}hlich and Vladimir Fuka and Katrin Frieda Gehrke and Jan Geleti{\v{c}} and Sebastian Giersch and Tobias Gronemeier and G\"{u}nter Gro{\ss} and Wieke Heldens and Antti Hellsten and Fabian Hoffmann and Atsushi Inagaki and Eckhard Kadasch and Farah Kanani-S\"{u}hring and Klaus Ketelsen and Basit Ali Khan and Christoph Knigge and Helge Knoop and Pavel Kr{\v{c}} and Mona Kurppa and Halim Maamari and Andreas Matzarakis and Matthias Mauder and Matthias Pallasch and Dirk Pavlik and Jens Pfafferott and Jaroslav Resler and Sascha Rissmann and Emmanuele Russo and Mohamed Salim and Michael Schrempf and Johannes Schwenkel and Gunther Seckmeyer and Sebastian Schubert and Matthias S\"{u}hring and Robert von Tils and Lukas Vollmer and Simon Ward and Bj\"{o}rn Witha and Hauke Wurps and Julian Zeidler and Siegfried Raasch},
+  title     = {Overview of the {PALM} model system 6.0},
+  journal   = {Geoscientific Model Development}
+}
+
+@article{Weller1998,
+  doi       = {10.1063/1.168744},
[…]
