+++ paper/paper.md
@@ -1,0 +1,63 @@
+---
+title: 'uDALES: large-eddy-simulation software for urban flow, dispersion, and microclimate modelling'
+tags:
+  - fluid dynamics
+  - computational fluid dynamics
+  - large eddy simulation
+  - urban climate
+  - fortran
+authors:
+  - name: Tom Grylls
+    orcid: 0000-0002-0948-4932
+    affiliation: 1
+  - name: Ivo Suter
+    orcid: 0000-0002-8612-4033
+    affiliation: "1, 2"
+  - name: Birgit S. Sützl
+    orcid: 0000-0001-9638-5643
+    affiliation: 1
+  - name: Sam Owens
+    orcid: 0000-0003-3253-1896
+    affiliation: 1
+  - name: David Meyer
+    orcid: 0000-0002-7071-7547
+    affiliation: "3, 1"
+  - name: Maarten van Reeuwijk
+    orcid: 0000-0003-4840-5050
+    affiliation: 1
+affiliations:
+ - name: Department of Civil and Environmental Engineering, Imperial College London, London, UK
+   index: 1
+ - name: Empa, Swiss Federal Laboratories for Materials Science and Technology, Dübendorf, Switzerland
+   index: 2
+ - name: Department of Meteorology, University of Reading, Reading, UK
+   index: 3
+date: 10 January 2021
+bibliography: paper.bib
+---
+
+
+# Summary
+
+With continuing urbanization, challenges associated with the urban environment such as air quality, heat islands, pedestrian thermal comfort, and wind loads on tall buildings, are increasingly relevant. Our ability to realistically capture processes such as the transport of heat, moisture, momentum and pollutants, and those of radiative transfer in urban environments is key to understanding and facing these challenges [@Oke2017]. The turbulent nature of the urban flow field and the inherent heterogeneity and wide range of scales associated with the urban environment result in a complex modelling problem. Large-eddy simulation (LES) is an approach to turbulence modelling used in computational fluid dynamics to simulate turbulent flows over a wide range of spatial and temporal scales. LES is one of the most promising tools to model the interactions typical of urban areas due to its ability to resolve the urban flow field at resolutions of $O$(1 m, 0.1 s), over spatial domains of $O$(100 m), and time periods of $O$(10 h). Although there are many scalable LES models for atmospheric flows, to our knowledge, only few are capable of explicitly representing buildings and of modelling the full range of urban processes (e.g. PALM-4U @Resler2017; @Maronga2020; or OpenFoam @Weller1998).
+
[…]
