### 2021-07-17 b13bdfe
Fixes to bottom edge wall function as identified in #23 (#135)
