+++ docs/udales-how-to-cite.md
@@ -1,0 +1,23 @@
+# How to cite
+
+When using the uDALES please cite as:
+
+> Grylls, T., Suter, I., Sützl, B., Owens, S., Meyer, D., & van Reeuwijk, M. (2021). uDALES: large-eddy-simulation software for urban flow, dispersion, and microclimate modelling. Journal of Open Source Software, 6(63), 3055, https://doi.org/10.21105/joss.03055
+
+
+```bibtex
+@article{Grylls2021,
+  doi = {10.21105/joss.03055},
+  url = {https://doi.org/10.21105/joss.03055},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {63},
+  pages = {3055},
+  author = {Tom Grylls and Ivo Suter and Birgit S. Sützl and Sam Owens and David Meyer and Maarten van Reeuwijk},
+  title = {uDALES: large-eddy-simulation software for urban flow, dispersion, and microclimate modelling},
+  journal = {Journal of Open Source Software}
+}
+```
+
+If you are looking to also cite a specific version of the software, you can find correposding version-specific DOIs on [Zenodo](10.5281/zenodo.5111496).
