+++ docs/udales-how-to-cite.md
@@ -20,4 +20,4 @@ When using the uDALES please cite as:
 }
 ```
 
-If you are looking to also cite a specific version of the software, you can find correposding version-specific DOIs on [Zenodo](10.5281/zenodo.5111496).
+If you are looking to also cite a specific version of the software, you can find correposding version-specific DOIs on [Zenodo](https://doi.org/10.5281/zenodo.5111496).
